Adding ability to choose enabled tab via the tab url param
Previously, after creating a new vlan, the user would be redirected to
the details page and the ports tab would be selected. This was
confusing to users. You may now pass tab=vlan|port|switch in the url
to specify which tab should be selected on load. By default the switch
tab is selected.