Replacing _process_interfaces with get_interfaces_state
The way we parsed output from `show interfaces brief` was broken. If
the port name had spaces, only the first word was recorded. In
addition, the second word was recorded as the type.

Using NETCONF and GET interface-statedata, we're getting the same data
minus trunk and type. Spaces in port names are no longer an issue, and
the port ID is now reported in full (ethernet 15/2 vs 15/2).