Adding polling for VLANs configured on each device
If a VLAN is modified switch side, then VCE should notice the
modification. This change polls the VLANs from the device via NetConf
and compares each VLANs' configured interfaces against VCE's network
model. If a difference is found, the state found on the device is
saved to VCE's network model.