Restoring ability to modify VLANs
When both adding and modifying a VLAN, check_tag_availability was
being used to validate the VLAN's parameters. Because an edit modifies
a VLAN that's already installed, this check caused edit requests to
fail.

I've moved check_tag_availbility from validate_circuit into
provision_vlan. This preserves the check for duplicate VLANs when
adding, and allows for edits to pass validation.