Adding error check to bail on first encountered error.
Based on feedback I've added error checking on each iteration of the
add and remove VLAN loops, but this does not address the larger issue
of rollback on failure. If a VLAN is removed from the first interface
but not the second, we should attempt to rollback the changes. An
issue has been created to track the more detailed fix.