Modifying Switch.(no_)interface_tagged to take an array of interfaces
Handling the failure case, when only a portion of interfaces are
successfully tagged, is more easily handled when all are sent to the
device at one time; This limits the number of failures that may occur
due to network or configuration errors.

This change also updates a few places where error handling was causing
a failure. This error was introduced when the response format between
Provisioning.pm and Switch.pm was changed.

Includes the ability to modify a VLAN's description.