Querying a device immediately after running a command of type 'action'
Because device state is queried on a 5m poll interval changes made
directly against the device may not be reflected to the user for a
long period.

This change calls `_gather_operational_status` after commands of type
'action' are executed. While this increases response time to the user,
it significantly reduces the time between when the command is made and
the change is reflected on the GUI.

Future work may reduce reponse time by limiting the scope of device
state polling or by asynchronously querying for device state.