Removing vlan cleanup from delete_interface
The original thought was that if an interface was deleted, vlans on
that interface should be cleaned up. But because we're using interface
names (not ids) in our vlan definitions, vlans on a multi-switch
deployment could be incorrectly modified.

Anyway, vlans should be sync'd correctly since vlans are updated
immediately after interface changes are synced.