Adding logic to issue_command to handle cli write commands
The last step in running a write command is to commit the config
changes via 'write mem'. When we last updated the issue_command method
to handle templated commands, we failed to handle this case which led
to uncommitted changes being added to the device. This also had the
effect of triggering a device error between cli command executions
`Cannot undo the configuration as ... session is using this mode`.

This change passes the command type '(read|write)' to issue_command
and issues 'write mem' before exiting the cli context switches.